Convert existing stuff to vuetify
