+++ src/web/views/LoginView.vue
@@ -10,10 +10,9 @@ const userData = ref({
     id: "",
 })
 
-
-const { mutate: registerMutation, onDone, error } = useMutation(gql`
-    mutation register($user: String!, $pass: String!) {
-        register(username: $user, password: $pass) {
+const { mutate: loginMutation, onDone, error } = useMutation(gql`
+    mutation login($user: String!, $pass: String!) {
+        login(username: $user, password: $pass) {
             id
             username
         }
@@ -21,11 +20,11 @@ const { mutate: registerMutation, onDone, error } = useMutation(gql`
 `)
 
 onDone((result) => {
-    userData.value = result.data.register
+    userData.value = result.data.login
 })
 
-function register() {
-    registerMutation({
+function login() {
+    loginMutation({
         user: username.value,
         pass: password.value,
     })
@@ -33,19 +32,19 @@ function register() {
 </script>
 
 <template>
-  <h1>Register</h1>
+  <h1>Login</h1>
+  <v-text-field
+    v-model="username"
+    label="Username"
+  />
+  <v-text-field
+    v-model="password"
+    label="Password"
+  />
   <p>
-    Username
-  </p>
-  <input v-model="username">
-  <p>
-    Password
-  </p>
-  <input v-model="password">
-  <p>
-    <button @click="register">
-      Register
-    </button>
+    <v-btn @click="login">
+      Login
+    </v-btn>
   </p>
   <p v-if="userData">
     {{ userData }}
